Add top section elements and styling
This commit adds just about everything that's in the upper part of the
website. I added:

-A nav element that contains a header logo (unsure if I will add a photo
 or not), as well as three links in the header (they probably won't lead
 anywhere)

-Below the header, there is a section that contains a heading, a
paragraph under the text, and a button (that will also lead nowhere)

-There is also a picutre, along with photo credits underneath

So far, I was correct in assuming that all of this can be done via
flexboxes. The header is a flexbox, and the orientation of the left and
right content is separated via two flexboxes. The header logo is its own
box, and the same goes for the header links.

Additionally, the text and button on the left side of the screen is
contained in one flexbox, and the picture with the photo credit is also
contained in a flexbox.

Lastly, I'm not too concerned with making sure that everything is spaced
correctly just yet, as I'll probably fine-tweak it later toward the end.

So far so good!