Add placeholder items in html
This commit intends to add all placeholders involved in the project
(various headings, text, buttons, footers, etc.) so that I can visualize
what I will be working with. Now that all the content is (hopefully)
there, and the HTML code is mostly there, I will be able to focus on css
for the most part. The only think that is not clearly there are the
boxes in the second block of the page. I am unsure how I will be
changing all palceholder items so far, but we'll see.

On to the hard part.